duplicated LoginActivity and activity_login
the login activity is a template generated from android studio. there is a lot of code in the file that i am not certain how to use.
i duplicated/renamed these files because i am going to overwrite them with my own code but want to keep the templates for reference/use in the future.