adds getUserName method to TaskListActivity
the getUserName method looks at the type field in the User object and
determines if user is manager or employee. if manager, the user name
for the task list is obtained from the bundle passed from the
NewHireListActivity. if the user is employee, the user's name is taken
from the User object itself